Adding this new class to the library.
Most of it's code come from bin file generateRRDGraphics.py
which got a major overhaul as to be able to cleany be able
to support multiple languages AT ONCE. This new class will 
help this program to produce graphics in a certain language while
the application might have been installed in another language.

One step close to having an application capable of generating 
multi-languages artifacts.