Added __setBasicPaths private method.
Added basicPaths concept which are the base paths of the application. These paths CANNOT
be translated and remain the same no matter hwat language is used. 

Amongst other things, this will keep us from having to deal with variable paths when 
dealing with imports.

Paths towards generated data( graphics, csv files, databases, webPages ) will
be translated in an effort to translates the application as much as is 
easily feasible.